+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.1.0
+
+### Minor Changes
+
+- logger added
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.4.0
+  - @volontariapp/errors@0.3.0
+
 ## 2.0.1
 
 ### Patch Changes