+++ packages/bridge-nest/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.2.0
+
+### Minor Changes
+
+- logger added
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.4.0
+  - @volontariapp/bridge@0.2.0
+  - @volontariapp/errors@0.3.0
+
 All notable changes to this project will be documented in this file.
 
 ## 0.1.0
@@ -7,4 +20,3 @@ All notable changes to this project will be documented in this file.
 ### Minor Changes
 
 - Initial package scaffold.
-